continue with release handling only if a release was found

+++ configuration/github-download-release-assets.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Download release tar ball and release-notes.txt and upload to releases.sapsailing.com
+# Download rlease tar ball and release-notes.txt and upload to releases.sapsailing.com
 # Usage:
 #   ./github-download-release-assets.sh {BEARER_TOKEN} {release-name-prefix}
 # For example:
@@ -18,23 +18,27 @@
 BEARER_TOKEN="${1}"
 RELEASE_NAME_PREFIX="${2}"
 RELEASES=$( curl -L -H 'Authorization: Bearer '${BEARER_TOKEN} https://api.github.com/repos/SAP/sailing-analytics/releases 2>/dev/null )
-RELEASE_NOTES_TXT_ASSET_ID=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="text/plain").id' )
-RELEASE_TAR_GZ_ASSET_ID=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="application/x-tar").id' )
-RELEASE_FULL_NAME=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="application/x-tar").name' | sed -e 's/\.tar\.gz$//')
-# For backward compatibility with deployed versions of java/target/refreshInstance.sh
-# which may be looking for a default release named "build-...", in case the release
-# is a "main-..." release, additionally create a corresponding folder and symbolic links named "build-..."
-RELEASE_NAME=$( echo ${RELEASE_FULL_NAME} | sed -e 's/^\(.*\)-\([0-9]*\)$/\1/' )
-RELEASE_TIMESTAMP=$( echo ${RELEASE_FULL_NAME} | sed -e 's/^\(.*\)-\([0-9]*\)$/\2/' )
-echo "Found release ${RELEASE_FULL_NAME} with name ${RELEASE_NAME} and time stamp ${RELEASE_TIMESTAMP}, notes ID is ${RELEASE_NOTES_TXT_ASSET_ID}, tarball ID is ${RELEASE_TAR_GZ_ASSET_ID}"
-curl -o ${RELEASE_FULL_NAME}.tar.gz -L -H 'Accept: application/octet-stream' -H 'Authorization: Bearer '${BEARER_TOKEN} 'https://api.github.com/repos/SAP/sailing-analytics/releases/assets/'${RELEASE_TAR_GZ_ASSET_ID}
-curl -o release-notes.txt -L -H 'Accept: application/octet-stream' -H 'Authorization: Bearer '${BEARER_TOKEN} 'https://api.github.com/repos/SAP/sailing-analytics/releases/assets/'${RELEASE_NOTES_TXT_ASSET_ID}
-FOLDER=releases/${RELEASE_FULL_NAME}
-ssh trac@sapsailing.com mkdir -p ${FOLDER}
-scp ${RELEASE_FULL_NAME}.tar.gz trac@sapsailing.com:${FOLDER}
-scp release-notes.txt trac@sapsailing.com:${FOLDER}
-if [ "${RELEASE_NAME}" = "main" ]; then
-  LINKING_FOLDER=releases/build-${RELEASE_TIMESTAMP}
-  ssh trac@sapsailing.com "mkdir -p ${LINKING_FOLDER}; ln -s ../${RELEASE_FULL_NAME}/${RELEASE_FULL_NAME}.tar.gz ${LINKING_FOLDER}/build-${RELEASE_TIMESTAMP}.tar.gz; ln -s ../${RELEASE_FULL_NAME}/release-notes.txt ${LINKING_FOLDER}"
+RELEASE_NOTES_TXT_ASSET_ID=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="text/plain").id' 2>/dev/null)
+if [ "$?" -ne "0" ]; then
+  echo "No release with prefix ${RELEASE_NAME_PREFIX} found. Not trying to download/upload anything."
+else
+  RELEASE_TAR_GZ_ASSET_ID=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="application/x-tar").id' )
+  RELEASE_FULL_NAME=$( echo "${RELEASES}" | jq -r 'sort_by(.created_at) | reverse | map(select(.name | startswith("'${RELEASE_NAME_PREFIX}'")))[0].assets[] | select(.content_type=="application/x-tar").name' | sed -e 's/\.tar\.gz$//')
+  # For backward compatibility with deployed versions of java/target/refreshInstance.sh
+  # which may be looking for a default release named "build-...", in case the release
+  # is a "main-..." release, additionally create a corresponding folder and symbolic links named "build-..."
+  RELEASE_NAME=$( echo ${RELEASE_FULL_NAME} | sed -e 's/^\(.*\)-\([0-9]*\)$/\1/' )
+  RELEASE_TIMESTAMP=$( echo ${RELEASE_FULL_NAME} | sed -e 's/^\(.*\)-\([0-9]*\)$/\2/' )
+  echo "Found release ${RELEASE_FULL_NAME} with name ${RELEASE_NAME} and time stamp ${RELEASE_TIMESTAMP}, notes ID is ${RELEASE_NOTES_TXT_ASSET_ID}, tarball ID is ${RELEASE_TAR_GZ_ASSET_ID}"
+  curl -o ${RELEASE_FULL_NAME}.tar.gz -L -H 'Accept: application/octet-stream' -H 'Authorization: Bearer '${BEARER_TOKEN} 'https://api.github.com/repos/SAP/sailing-analytics/releases/assets/'${RELEASE_TAR_GZ_ASSET_ID}
+  curl -o release-notes.txt -L -H 'Accept: application/octet-stream' -H 'Authorization: Bearer '${BEARER_TOKEN} 'https://api.github.com/repos/SAP/sailing-analytics/releases/assets/'${RELEASE_NOTES_TXT_ASSET_ID}
+  FOLDER=releases/${RELEASE_FULL_NAME}
+  ssh trac@sapsailing.com mkdir -p ${FOLDER}
+  scp ${RELEASE_FULL_NAME}.tar.gz trac@sapsailing.com:${FOLDER}
+  scp release-notes.txt trac@sapsailing.com:${FOLDER}
+  if [ "${RELEASE_NAME}" = "main" ]; then
+    LINKING_FOLDER=releases/build-${RELEASE_TIMESTAMP}
+    ssh trac@sapsailing.com "mkdir -p ${LINKING_FOLDER}; ln -s ../${RELEASE_FULL_NAME}/${RELEASE_FULL_NAME}.tar.gz ${LINKING_FOLDER}/build-${RELEASE_TIMESTAMP}.tar.gz; ln -s ../${RELEASE_FULL_NAME}/release-notes.txt ${LINKING_FOLDER}"
+  fi
+  rm ${RELEASE_FULL_NAME}.tar.gz release-notes.txt
 fi
-rm ${RELEASE_FULL_NAME}.tar.gz release-notes.txt
